event_management/requirements.txt
django~=3.2.8
djangorestframework~=3.12.4
drf-yasg~=1.20.0  # For Swagger/OpenAPI documentation generation
